Animations

+++ Version_LATEST/src/main/java/com/acrylic/version_latest/Animations/AbstractAnimations.java
@@ -71,9 +71,7 @@ public abstract class AbstractAnimations {
 
     public void delete() {
         entity.remove();
-        if (holograms != null) {
-            holograms.delete();
-        }
+        if (holograms != null) holograms.delete();
     }
 
     public abstract void teleport(Location location);
